requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
httpx==0.26.0
huggingface_hub==0.20.3
